+++ DummyDevice.h
@@ -20,8 +20,8 @@ public:
 	virtual void Initialize();
 	virtual void Terminate();
 	//for real device
-	virtual void Write(unsigned long long value, unsigned block, unsigned byte_offset, unsigned bit_offset, unsigned bits);
-	virtual unsigned long long Read(unsigned block, unsigned byte_offset, unsigned bit_offset, unsigned bits);
+	virtual void Write(unsigned int value, unsigned block, unsigned byte_offset, unsigned bit_offset, unsigned bits);
+	virtual unsigned int Read(unsigned block, unsigned byte_offset, unsigned bit_offset, unsigned bits);
 };
 
 #endif /* DUMMYDEVICE_H_ */
